Move DefaultFormatter, NumberFormatter and OptionalNumberFormatter to formatters.js, where they belong
