+++ src-js/fontra-webcomponents/src/ui-list.js
@@ -1,6 +1,6 @@
+import { DefaultFormatter } from "@fontra/core/formatters.js";
 import * as html from "@fontra/core/html-utils.js";
 import { UnlitElement } from "@fontra/core/html-utils.js";
-import { DefaultFormatter } from "@fontra/core/ui-utils.js";
 import { message } from "@fontra/web-components/modal-dialog.js";
 import { themeColorCSS } from "./theme-support.js";
